requirements.txt
fastapi
uvicorn
openai
langchain
chromadb
pytesseract
Pillow
python-dotenv
requests
